Bug in timestamp retrieval function for reading a video file, it was being cast to a uint16_t instead of a uint64_t

+++ src/Libraries/VideoSource/FileVideoSource.cpp
@@ -59,12 +59,13 @@ int FileVideoSource::CurrentFrameNo()
 {
 	return m_videoFile->get(CV_CAP_PROP_POS_FRAMES);
 }
+
 uint64_t FileVideoSource::GetTimestamp()
 {
 	// returns timestamp in milliseconds appended with "Frame_"
 	//std::string frame_timestamp;
 	//frame_timestamp="Frame_"+std::to_string(m_videoFile->get(CV_CAP_PROP_POS_MSEC));
-	return (uint16_t)m_videoFile->get(CV_CAP_PROP_POS_MSEC);
+	return (uint64_t)m_videoFile->get(CV_CAP_PROP_POS_MSEC);
 }
 
 }
